+++ testhiml/health/azure/test_download_aml_run.py
@@ -32,10 +32,10 @@ def test_no_config_path() -> None:
         assert "You must provide a config.json file in the root folder to connect" in str(e)
 
 
-def test_download_aml_run_no_runs() -> None:
+def test_download_aml_run_no_runs(tmp_path: Path) -> None:
     # if no such run exists, will fail
     with pytest.raises(Exception) as e:
-        subprocess.Popen(["python", DOWNLOAD_SCRIPT_PATH, "--run_id", "madeuprun"])
+        subprocess.Popen(["python", DOWNLOAD_SCRIPT_PATH, "--run_id", "madeuprun" "--output_dir", str(tmp_path)])
         assert "was not found" in str(e)